+++ CHANGELOG.md
@@ -70,3 +70,4 @@
 - Bug #66: Refactor test names and messages in `HasLangTest` and `LangProvider` for clarity (@terabytesoftw)
 - Bug #67: Improve `sanitizeJsonValue` method in `BaseAttributes` class to support callable values and update documentation for clarity (@terabytesoftw)
 - Bug #68: Update class reference in docblock for `EncodeProvider` to improve clarity and accuracy (@terabytesoftw)
+- Bug #69: Add test case for closure returning `float` value in `AttributesProvider` class (@terabytesoftw)